adding comments management for users, view posts base on username on views

+++ BetterBlog/controlPanel.py
@@ -71,10 +71,11 @@ def delete_user():
 @manage.route('/posts', methods=["GET", "POST"])
 @login_required
 def posts():
-    if request.method == "GET":
-        page = request.args.get('page', 1, type=int)
-        pagination = Post.query.filter_by(author=current_user.username).paginate(page=page, per_page=2)
-        return render_template("admin/posts.html", posts=pagination, pagination=pagination)
+    if current_user.is_Author:
+        if request.method == "GET":
+            page = request.args.get('page', 1, type=int)
+            pagination = Post.query.filter_by(author=current_user.username).paginate(page=page, per_page=2)
+            return render_template("admin/posts.html", posts=pagination, pagination=pagination)
 
 
 # delete post
@@ -85,9 +86,13 @@ def delete_post():
         post_id = request.form["delete_post"]
         post = db.session.query(Post).filter_by(
             id=post_id).first()
-        db.session.delete(post)
-        db.session.commit()
-        flash("Delete Success!",category="success")
+        if post.author == current_user.username:
+            db.session.delete(post)
+            db.session.commit()
+            flash("Delete Success!", category="success")
+        else:
+            flash("You are not allow to edit this post!",category="danger")
+
         return redirect(url_for("controlPanel.posts"))
     except Exception:
         return render_template("handler/404.html")
@@ -99,29 +104,28 @@ def delete_post():
 def edit_post(post_id):
     post = db.session.query(Post).filter_by(
         id=post_id).first()  # use the object itself instead of the copy version of it
-
-    edit_form = CreatePostForm(  # create as a form to able to get the old data to the wtflask form (faster way)
-        title=post.title,
-        subtitle=post.subtitle,
-        img_url=post.img_url,
-        author=post.author,
-        body=post.body
-    )
-    if edit_form.validate_on_submit():
-        post.title = edit_form.title.data
-        post.subtitle = edit_form.subtitle.data
-        post.img_url = edit_form.img_url.data
-        post.author = edit_form.author.data
-        post.body = strip_invalid_html(edit_form.body.data)
-        print("updating....")
-        try:
-            print(edit_form)
-            db.session.commit()
-        except SQLAlchemyError:
-            print("can't update, something is wrong while editing!")
-            db.session.rollback()
-            raise
-        return redirect(url_for("views.post", post_id=post.id))
+    if post.author == current_user.username:
+        edit_form = CreatePostForm(  # create as a form to able to get the old data to the wtflask form (faster way)
+            title=post.title,
+            subtitle=post.subtitle,
+            img_url=post.img_url,
+            author=post.author,
+            body=post.body
+        )
+        if edit_form.validate_on_submit():
+            post.title = edit_form.title.data
+            post.subtitle = edit_form.subtitle.data
+            post.img_url = edit_form.img_url.data
+            post.author = edit_form.author.data
+            post.body = strip_invalid_html(edit_form.body.data)
+            print("updating....")
+            try:
+                db.session.commit()
+            except SQLAlchemyError:
+                print("can't update, something is wrong while editing!")
+                db.session.rollback()
+                raise
+            return redirect(url_for("views.post", post_id=post.id))
         # return redirect(url_for('controlPanel.posts'))
-
-    return render_template("handler/createpost.html", form=edit_form, is_edit=True, name=current_user.username)
+        return render_template("handler/createpost.html", form=edit_form, is_edit=True, name=current_user.username)
+    return redirect(url_for("views.get_all_posts"))
